requirements.txt
huggingface_hub[cli,hf_transfer]
pyyaml